Moved 'isCopyEliminated' function to CopyEliminationAnalyzer
